tests/requirements.txt
pyserial
adafruit-circuitpython-pn532
